これまでのコマンドをdebugコマンドとし、サブコマンドなしのコマンドを追加
